% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fishSim_0.11.R
\name{birthdays}
\alias{birthdays}
\title{birthdays(): takes an individual-data matrix, and adds one to each individual's age}
\usage{
birthdays(indiv = makeFounders())
}
\arguments{
\item{indiv}{A matrix of individuals, as from makeFounders(), move(), mate(), or mort().}
}
\description{
Nothing fancy: this is separated from the other functions to allow more flexible
assignments of movement, mating and mortality within each year.
}
